Python/requirements.txt
configparser==3.5.0
python-telegram-bot==5.0.0
redis==2.10.5